# docs/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Ninjo API
  version: "1.0.0"
  description: 勉強計画を立て、AIと一緒に「やり切る」を実現するAPI
servers:
  - url: /api/v1
paths:
...
  /tasks/{task_id}/extract-skills:
    post:
      operationId: extractSkills
      summary: タスクのメモからAIがスキルを抽出（保存しない）
      tags: [skill]
      security:
        - cookieAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 抽出成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractSkillsResponse"
        "400":
          description: メモが空
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: タスクが存在しない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /skills:
    get:
      operationId: listSkills
      summary: ユーザーの全スキル一覧を取得
      tags: [skill]
      security:
        - cookieAuth: []
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: カテゴリでフィルタ
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillListResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createSkill
      summary: スキルを手動追加
      tags: [skill]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkillRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 同名スキルが既に存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /skills/{skill_id}:
    put:
      operationId: updateSkill
      summary: スキルの名前やカテゴリを更新
      tags: [skill]
      security:
        - cookieAuth: []
      parameters:
        - name: skill_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSkillRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: スキルが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSkill
      summary: スキルを削除
      tags: [skill]
      security:
        - cookieAuth: []
      parameters:
        - name: skill_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 削除成功
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: スキルが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ExtractSkillsResponse:
      type: object
      required: [suggested_skills]
      properties:
        suggested_skills:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedSkill"
    SuggestedSkill:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        category:
          type: string
    CreateSkillRequest:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        category:
          type: string
          minLength: 1
          maxLength: 100
        task_id:
          type: string
          format: uuid
          description: タスクに紐付ける場合のみ指定
    UpdateSkillRequest:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        category:
          type: string
          minLength: 1
          maxLength: 100
    SkillResponse:
      type: object
      required: [id, name, category, source, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        source:
          type: string
          enum: [ai, manual]
        task_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    SkillListResponse:
      type: object
      required: [skills, total_count]
      properties:
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillResponse"
        total_count:
          type: integer
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token